:hammer: [chore] StoryTableView 추가 #12
